+++ lib/govuk_app_config/govuk_i18n.rb
@@ -3,6 +3,19 @@ module GovukI18n
     {
       # Azerbaijani
       az: { i18n: { plural: { keys: %i[one other], rule: ->(n) { n == 1 ? :one : :other } } } },
+      # Welsh
+      cy: { i18n: { plural: { keys: %i[zero one two few many other],
+                              rule:
+                                lambda do |n|
+                                  case n
+                                  when 0 then :zero
+                                  when 1 then :one
+                                  when 2 then :two
+                                  when 3 then :few
+                                  when 6 then :many
+                                  else :other
+                                  end
+                                end } } },
       # Dari - this isn't an iso code. Probably should be 'prs' as per ISO 639-3.
       dr: { i18n: { plural: { keys: %i[one other], rule: ->(n) { n == 1 ? :one : :other } } } },
       # Latin America and Caribbean Spanish
@@ -31,12 +44,31 @@ module GovukI18n
       kk: { i18n: { plural: { keys: %i[one other], rule: ->(n) { n == 1 ? :one : :other } } } },
       # Punjabi Shahmukhi
       "pa-pk": { i18n: { plural: { keys: %i[one other], rule: ->(n) { n == 1 ? :one : :other } } } },
+      # Maltese
+      mt: { i18n: { plural: { keys: %i[one few many other],
+                              rule:
+                                lambda do |n|
+                                  n ||= 0
+                                  mod100 = n % 100
+
+                                  if n == 1
+                                    :one
+                                  elsif n.zero? || (2..10).to_a.include?(mod100)
+                                    :few
+                                  elsif (11..19).to_a.include?(mod100)
+                                    :many
+                                  else
+                                    :other
+                                  end
+                                end } } },
       # Sinhalese
       si: { i18n: { plural: { keys: %i[one other], rule: ->(n) { n == 1 ? :one : :other } } } },
       # Turkish
       tr: { i18n: { plural: { keys: %i[one other], rule: ->(n) { n == 1 ? :one : :other } } } },
       # Uzbek
       uz: { i18n: { plural: { keys: %i[one other], rule: ->(n) { n == 1 ? :one : :other } } } },
+      # Chinese
+      zh: { i18n: { plural: { keys: %i[other], rule: ->(_) { :other } } } },
       # Chinese Hong Kong
       "zh-hk" => { i18n: { plural: { keys: %i[one other], rule: ->(n) { n == 1 ? :one : :other } } } },
       # Chinese Taiwan